rename nav-link, radio, text-area, text-input, text-link
