+++ packages/radio/src/Radio.stories.tsx
@@ -5,7 +5,7 @@ import { Stack } from '@spark-web/stack';
 import { Text } from '@spark-web/text';
 import type { ComponentMeta, ComponentStory } from '@storybook/react';
 
-import { Radio } from './radio';
+import { Radio } from './Radio';
 
 export default {
   title: 'Forms / Radio',
