+++ packages/text-area/src/index.ts
@@ -1,5 +1,5 @@
-export { TextArea } from './text-area';
+export { TextArea } from './Text-Area';
 
 // types
 
-export type { TextAreaProps } from './text-area';
+export type { TextAreaProps } from './Text-Area';
