+++ packages/text-input/src/index.ts
@@ -1,7 +1,7 @@
 export { InputAdornment } from './InputAdornment';
-export { TextInput, useInput } from './TextInput';
+export { TextInput, useInput } from './Text-Input';
 
 // types
 
 export type { AdornmentChild } from './childrenToAdornments';
-export type { TextInputProps, UseInputProps } from './TextInput';
+export type { TextInputProps, UseInputProps } from './Text-Input';
