+++ packages/text-link/src/index.ts
@@ -1,7 +1,7 @@
-export { TextLink } from './TextLink';
+export { TextLink } from './Text-Link';
 export { TextLinkButton } from './TextLinkButton';
 
 // types
 
-export type { TextLinkProps } from './TextLink';
+export type { TextLinkProps } from './Text-Link';
 export type { TextLinkButtonProps } from './TextLinkButton';
